UI: Fix compiler warnings
CMake build framework 3.0 introduces more strict typing checks, which
require additional code fixes to silence.

* Fixes warnings about implicit casts
* Fixes warnings about shadowing variables
* Fixes warnings about unused assignments

+++ UI/audio-encoders.cpp
@@ -78,8 +78,8 @@ static void HandleSampleRate(obs_property_t *prop, const char *id)
 		return;
 	}
 
-	uint32_t sampleRate =
-		config_get_uint(main->Config(), "Audio", "SampleRate");
+	uint32_t sampleRate = (uint32_t)config_get_uint(main->Config(), "Audio",
+							"SampleRate");
 
 	obs_data_set_int(data.get(), "samplerate", sampleRate);
