+++ UI/display-helpers.hpp
@@ -48,8 +48,8 @@ static inline void GetCenterPosFromFixedScale(int baseCX, int baseCY,
 					      int windowCX, int windowCY,
 					      int &x, int &y, float scale)
 {
-	x = (float(windowCX) - float(baseCX) * scale) / 2.0f;
-	y = (float(windowCY) - float(baseCY) * scale) / 2.0f;
+	x = (int)((float(windowCX) - float(baseCX) * scale) / 2.0f);
+	y = (int)((float(windowCY) - float(baseCY) * scale) / 2.0f);
 }
 
 static inline QSize GetPixelSize(QWidget *widget)
